+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Improved username validation.
+- Hide subscription settings page if user doesn't have permission to manage subscriptions.
 
 ### Fixed